+++ packages/core/select/src/createSelect.js
@@ -1,6 +1,6 @@
 // @flow
 import React, { Component, type ComponentType, type ElementRef } from 'react';
-import { mergeStyles } from 'react-select';
+import { mergeStyles, makeAnimated } from 'react-select';
 import memoizeOne from 'memoize-one';
 import isEqual from 'react-fast-compare';
 import { colors, gridSize } from '@atlaskit/theme';
@@ -266,10 +266,10 @@ export default function createSelect(WrappedComponent: ComponentType<*>) {
       this.cacheComponents(nextProps.components);
     }
     cacheComponents = (components?: {}) => {
-      this.components = {
+      this.components = makeAnimated({
         ...defaultComponents,
         ...components,
-      };
+      });
     };
     focus() {
       this.select.focus();