+++ packages/editor/editor-core/src/__tests__/integration/card/inline-2.ts
@@ -25,7 +25,7 @@ const moveUp = (page, selector) => {
 BrowserTestCase(
   `inline-2.ts: pasting an link then typing still converts to inline card`,
   {
-    skip: ['ie', 'safari'],
+    skip: ['ie', 'safari', 'firefox'],
   },
   async client => {
     let browser = new Page(client);