+++ packages/editor/editor-core/src/__tests__/integration/panel/paste-plain-text.ts
@@ -16,7 +16,7 @@ import { selectors } from './_utils';
 
 BrowserTestCase(
   'paste-plain-text.ts: Paste plain text into panel',
-  { skip: ['ie', 'edge', 'safari'] },
+  { skip: ['ie', 'edge', 'safari', 'firefox'] },
   async client => {
     const page = await goToEditorTestingExample(client);