+++ packages/editor/editor-core/src/__tests__/integration/table/auto-size.ts
@@ -1,4 +1,5 @@
 import { BrowserTestCase } from '@atlaskit/webdriver-runner/runner';
+import { sleep } from '@atlaskit/editor-test-helpers';
 
 import { editable, getDocFromElement, fullpage } from '../_helpers';
 
@@ -30,13 +31,14 @@ async function loadAndRetrieveDocument(
   });
 
   await page.waitForSelector(`table[data-layout="${expectedLayout}"]`);
+  await sleep(500);
 
   const doc = await page.$eval(editable, getDocFromElement);
   return doc;
 }
 
 BrowserTestCase(
-  'auto-size.ts: Doesnt scale past default',
+  'Doesnt scale past default',
   { skip: ['ie', 'edge', 'safari', 'firefox'] },
   async client => {
     const page = await goToEditorTestingExample(client);
@@ -46,7 +48,7 @@ BrowserTestCase(
 );
 
 BrowserTestCase(
-  'auto-size.ts: Scales to wide',
+  'Scales to wide',
   { skip: ['ie', 'edge', 'safari', 'firefox'] },
   async client => {
     const page = await goToEditorTestingExample(client);
@@ -60,7 +62,7 @@ BrowserTestCase(
 );
 
 BrowserTestCase(
-  'auto-size.ts: Scales to full-width',
+  'Scales to full-width',
   { skip: ['ie', 'edge', 'safari', 'firefox'] },
   async client => {
     const page = await goToEditorTestingExample(client);