+++ packages/editor/editor-core/src/plugins/gap-cursor/actions.ts
@@ -149,6 +149,11 @@ export const setGapCursorAtPos = (
   position: number,
   side: Side = Side.LEFT,
 ): Command => (state, dispatch) => {
+  // @see ED-6231
+  if (position > state.doc.content.size) {
+    return false;
+  }
+
   const $pos = state.doc.resolve(position);
 
   if (GapCursorSelection.valid($pos)) {
@@ -166,9 +171,10 @@ export const setGapCursorAtPos = (
 const captureCursorCoords = (
   event: MouseEvent,
   editorRef: HTMLElement,
-  posAtCoords: (
-    coords: { left: number; top: number },
-  ) => { pos: number; inside: number } | null | void,
+  posAtCoords: (coords: {
+    left: number;
+    top: number;
+  }) => { pos: number; inside: number } | null | void,
   state: EditorState,
 ): { position: number; side: Side } | null => {
   const rect = editorRef.getBoundingClientRect();
@@ -209,9 +215,10 @@ const captureCursorCoords = (
 export const setCursorForTopLevelBlocks = (
   event: MouseEvent,
   editorRef: HTMLElement,
-  posAtCoords: (
-    coords: { left: number; top: number },
-  ) => { pos: number; inside: number } | null | void,
+  posAtCoords: (coords: {
+    left: number;
+    top: number;
+  }) => { pos: number; inside: number } | null | void,
 ): Command => (state, dispatch) => {
   // plugin is disabled
   if (!pluginKey.get(state)) {