+++ packages/elements/mention/src/__tests__/unit/api/MentionResourceSpec.ts
@@ -1,13 +1,12 @@
+import { SecurityOptions } from '@atlaskit/util-service-support';
 import 'es6-promise/auto'; // 'whatwg-fetch' needs a Promise polyfill
 import 'whatwg-fetch';
 import * as fetchMock from 'fetch-mock/src/client';
-import { SecurityOptions } from '@atlaskit/util-service-support';
 import * as queryString from 'query-string';
-
-import { MentionDescription } from '../../../types';
 import MentionResource, {
   MentionResourceConfig,
 } from '../../../api/MentionResource';
+import { MentionDescription } from '../../../types';
 import {
   resultC,
   resultCr,
@@ -29,7 +28,8 @@ const options = (
   omitCredentials,
 });
 
-const getSecurityHeader = call => call[1].headers.get(defaultSecurityHeader);
+const getSecurityHeader = (call: any) =>
+  call[1].headers.get(defaultSecurityHeader);
 
 const defaultSecurityCode = '10804';
 
@@ -47,7 +47,10 @@ const apiConfigWithoutCredentials: MentionResourceConfig = {
   },
 };
 
-function checkOrder(expected, actual) {
+function checkOrder(
+  expected: MentionDescription[][],
+  actual: MentionDescription[][],
+) {
   expect(actual).toHaveLength(expected.length);
   for (let i = 0; i < expected.length; i++) {
     expect(actual[i]).toHaveLength(expected[i].length);