+++ packages/elements/util-data-test/src/mention/presence-data.ts
@@ -2,8 +2,25 @@ declare var require: {
   <T>(path: string): T;
 };
 
-// tslint:disable-next-line:no-var-requires
-export const validPresenceData: Response = require('../json-data/presence-valid-info.json') as Response;
+export interface PresenceValidInfo {
+  data: Data;
+}
+
+export interface Data {
+  PresenceBulk: PresenceBulk[];
+}
+
+export interface PresenceBulk {
+  userId: string;
+  state: null | string;
+  type: null | string;
+  date: null | string;
+  message: null | string;
+  stateMetadata?: string;
+}
 
 // tslint:disable-next-line:no-var-requires
-export const invalidPresenceData: Response = require('../json-data/presence-invalid-info.json') as Response;
+export const validPresenceData: PresenceValidInfo = require('../json-data/presence-valid-info.json') as PresenceValidInfo;
+
+// tslint:disable-next-line:no-var-requires
+export const invalidPresenceData: PresenceValidInfo = require('../json-data/presence-invalid-info.json') as PresenceValidInfo;