+++ packages/media/media-card/example-helpers/index.tsx
@@ -1,7 +1,7 @@
 /* tslint:disable:no-console */
 
 import * as React from 'react';
-import { Context, MediaItem } from '@atlaskit/media-core';
+import { Context, FileItem } from '@atlaskit/media-core';
 import { createStorybookContext } from '@atlaskit/media-test-helpers';
 import CrossIcon from '@atlaskit/icon/glyph/cross';
 import AnnotateIcon from '@atlaskit/icon/glyph/media-services/annotate';
@@ -106,14 +106,14 @@ export const actions = [
 
 export const anotherAction: CardAction = {
   label: 'Some other action',
-  handler: (item?: MediaItem) => {
+  handler: (item?: FileItem) => {
     console.log('Some other action', item);
   },
 };
 
 export const annotateAction: CardAction = {
   label: 'Annotate',
-  handler: (item?: MediaItem) => {
+  handler: (item?: FileItem) => {
     console.log('annotate', item);
   },
 };