+++ packages/media/media-card/src/root/domain.ts
@@ -1,26 +1,10 @@
-export type Identifier =
-  | UrlPreviewIdentifier
-  | LinkIdentifier
-  | FileIdentifier
-  | ExternalImageIdentifier;
+export type Identifier = FileIdentifier | ExternalImageIdentifier;
 
 export interface FileIdentifier {
   readonly mediaItemType: 'file';
   readonly id: string | Promise<string>;
   readonly occurrenceKey?: string;
   readonly collectionName?: string; // files can exist outside of a collection
-}
-
-export interface LinkIdentifier {
-  readonly mediaItemType: 'link';
-  readonly id: string;
-  readonly occurrenceKey?: string;
-  readonly collectionName: string; // links always exist within a collection
-}
-
-export interface UrlPreviewIdentifier {
-  readonly mediaItemType: 'link';
-  readonly url: string;
 }
 
 export interface ExternalImageIdentifier {